// src/app/model/salary.model.ts
export class salaryModel {
    personalno?: number;
    name?: string;
    bank?: string;
    bankaccno? : number;
    dob?: number;
    lopdays?: number;
    pfno?: number;
    stddays?: number;
    location?: string;
    workingdays?: number;
    department?: string;
    designation?: string;
    basic?: number;
    pf?: number;
    houserent?: number;
    tax?: number;
    specialallowance?: number;
    incometax?: number;
    conveyance?: number;
    ADD?: number;
    shift?: number;
    bonus?: number;
    medical?: number;
    grossearn?: number;
    grossdeduction?:number;
    netpay?: number;
}